requirements.txt
python-dateutil
requests
selectorlib
playwright